pyproject.toml
[project]
name = "zndraw-socketio"
version = "0.1.1"
description = "Typed wrapper for python-socketio with Pydantic validation and dependency injection."
readme = "README.md"
authors = [
    { name = "Atomie CHEN", email = "atomic_cwh@163.com" },
    { name = "Fabian Zills", email = "fabian.zills@web.de" }
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = ["socketio", "pydantic", "wrapper", "typed", "zndraw"]
dependencies = [
    "pydantic>=2.10.6",
    "python-socketio>=5.12.1",
    "typing_extensions>=4.13.0",
]

[project.urls]
homepage = "https://github.com/pythonFZ/zndraw-socketio"
issues = "https://github.com/pythonFZ/zndraw-socketio/issues"

[project.optional-dependencies]
client = [
    "python-socketio[client]>=5.12.1",
]
asyncio-client = [
    "python-socketio[asyncio-client]>=5.12.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "aiohttp>=3.11.13",
    "fastapi[standard]>=0.115.11",
    "pyright>=1.1.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=5.0.0",
    "python-socketio[client]>=5.12.1",
    "ruff>=0.11.0",
    "typing_extensions>=4.13.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
